requirements.txt
Django>=4.0,<5.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
mysqlclient>=2.1.1
gunicorn>=20.1.0
